Can now edit, save as draft, publish etc, some other css tweaks
Still some bugs such as not logged in users can view drafts by changint he id param in url. When editing there is a slight bug in which the content does not show in the simplemde editor unless you click on it. The content is still being loaded it just does not display - more of a visual issue.